requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
datasets==1.17.0
dill==0.3.4
filelock==3.4.1
flatbuffers==2.0
fonttools==4.28.1
frozenlist==1.2.0
fsspec==2021.11.1
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gTTS==2.2.3
h5py==3.6.0
happytransformer==2.4.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.0
mediapipe==0.8.9
multidict==5.2.0
multiprocess==0.70.12.2
nltk==3.6.6
numpy==1.21.4
oauthlib==3.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
playsound==1.3.0
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
sentencepiece==0.1.96
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
sounddevice==0.4.3
SpeechRecognition==3.8.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tokenizers==0.10.3
tomli==1.2.2
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
typing-extensions==4.0.1
urllib3==1.26.7
wavio==0.0.4
Werkzeug==2.0.2
flask-socketio==5.3.4
python-socketio==5.11.0
python-engineio==4.7.1
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
